Complex types - mixed lists

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
 
 
-from typing import List, Union
+from typing import List
 
 
-def sum_mixed_list(mxd_lst: List[Union[float, int]]) -> float:
-    return sum(mxd_lst)
+def sum_mixed_list(mxd_lst: List[Union[int, float]]) -> float:
+    return float(sum(mxd_lst))
